fixed #122
