+++ ptmd/database/queries/files.py
@@ -76,12 +76,14 @@ def create_files(files_data: dict) -> list[File]:
     return files
 
 
-def get_shipped_file(species: str, batch: str) -> File | None:
+def get_shipped_file(species: str, batch: str) -> File | None:  # pragma: no cover
     """ This function return the first received file with the given batch and species if it exists or None
 
     :param species: the species of the file
     :param batch: the batch of the file
     :return: the first received file with the given batch and species if it exists or None
+
+    @note: This function is excluded from coverage report because it is just a wrapper around a database query
     """
     return File.query.join(Organism).filter(
             Organism.ptox_biosystem_name == species,
